Clarify build instruction + Touching up
Due to some difficulty I had while compiling because I changed the build configuration to `x64` which resulted in 30 minutes of time attempting to fix the build. Thanks to GitHub user ejbtrd's fork of the repository specifically commit 0cf42cd. 

I also attempted to touch up the README to make it a bit more appealing to the reader.